moved logout endpoint to login api
