+++ web/web/static/js/network_client.js
@@ -29,6 +29,17 @@ export const api = {
         }
     },
 
+    delete: async (resource, data, successHandler, errorHandler) => {
+        try {
+            const response = await api.network.delete(`${api.url}${resource}`, data).json();
+            if(successHandler && typeof successHandler === "function") {
+                successHandler(response);
+            }
+        } catch(e) {            
+            auth.handleError(e, errorHandler);
+        }
+    },
+
     handleApiError: (exception, errorHandler) => {
         if(exception instanceof ky.HTTPError) {
             console.warn("there was an HTTPError");
@@ -46,9 +57,9 @@ export const auth = {
     network: ky,    
     url: '/auth/',
 
-    delete: async (resource, data, successHandler, errorHandler) => {
+    post: async (resource, data, successHandler, errorHandler) => {
         try {
-            const response = await auth.network.delete(`${auth.url}${resource}`, data).json();
+            const response = await auth.network.post(`${auth.url}${resource}`, data).json();
             if(successHandler && typeof successHandler === "function") {
                 successHandler(response);
             }
